// packages/target-ethers-v6-test/types/v0.6.4/Events/Events.ts
export declare namespace Events {
  export type EventDataStruct = { index: BigNumberish; name: string };

  export type EventDataStructOutput = [index: bigint, name: string] & {
    index: bigint;
    name: string;
  };
}
// ...
  export type InputTuple = [value1: BigNumberish];
  export type OutputTuple = [value1: bigint];
  export interface OutputObject {
    value1: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
// ...
export namespace Event1Event {
  export type InputTuple = [value1: BigNumberish, value2: BigNumberish];
  export type OutputTuple = [value1: bigint, value2: bigint];
  export interface OutputObject {
    value1: bigint;
    value2: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}
// ...
export namespace Event3_bool_uint256_Event {
  export type InputTuple = [value1: boolean, value2: BigNumberish];
  export type OutputTuple = [value1: boolean, value2: bigint];
  export interface OutputObject {
    value1: boolean;
    value2: bigint;
  }
  export type Event = TypedContractEvent<InputTuple, OutputTuple, OutputObject>;
  export type Filter = TypedDeferredTopicFilter<Event>;
  export type Log = TypedEventLog<Event>;
  export type LogDescription = TypedLogDescription<Event>;
}